# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a golyadmin Server",
        "title": "Swagger API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/base/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Base"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户名, 密码, 验证码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequestParams"
                        }
                    }
                ],
                "responses": {
                    "2000": {
                        "description": "返回包括用户信息,token,过期时间",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/system.LoginResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/system/user/change_password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户修改密码",
                "parameters": [
                    {
                        "description": "用户名, 原密码, 新密码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ChangePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户修改密码",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/user/user": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户名, 密码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateUserRequestParams"
                        }
                    }
                ],
                "responses": {
                    "2000": {
                        "description": "创建用户,返回包括用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/request.CreateUserRequestParams"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.ChangePasswordReq": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "description": "新密码",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "request.CreateUserRequestParams": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "头像"
                },
                "email": {
                    "type": "string",
                    "example": "电子邮箱"
                },
                "gender": {
                    "type": "string",
                    "example": "性别"
                },
                "mobile": {
                    "type": "string",
                    "example": "电话号码"
                },
                "nickname": {
                    "type": "string",
                    "example": "昵称"
                },
                "password": {
                    "type": "string",
                    "example": "密码"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "request.LoginRequestParams": {
            "type": "object",
            "required": [
                "captcha",
                "password",
                "username"
            ],
            "properties": {
                "captcha": {
                    "description": "验证码",
                    "type": "string"
                },
                "captchaKey": {
                    "description": "验证码key",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "response.StructResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "system.LoginResponse": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/system.LyadminUsers"
                }
            }
        },
        "system.LyadminDept": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                }
            }
        },
        "system.LyadminMenu": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "component": {
                    "description": "对应前端文件路径",
                    "type": "string"
                },
                "component_name": {
                    "description": "对应前端文件名称",
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "is_catalog": {
                    "description": "是否目录",
                    "type": "boolean"
                },
                "is_link": {
                    "description": "是否外链",
                    "type": "boolean"
                },
                "keep_alive": {
                    "description": "是否缓存页面",
                    "type": "boolean"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "sort": {
                    "description": "显示顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                },
                "visible": {
                    "description": "是否显示菜单",
                    "type": "boolean"
                },
                "web_path": {
                    "description": "路由地址",
                    "type": "string"
                }
            }
        },
        "system.LyadminMenuButton": {
            "type": "object",
            "properties": {
                "api": {
                    "type": "string"
                },
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "menu_id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "system.LyadminPost": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                }
            }
        },
        "system.LyadminRole": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "dept": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminDept"
                    }
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "menu": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "permission": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminMenuButton"
                    }
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                }
            }
        },
        "system.LyadminUsers": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "dept": {
                    "description": "部门外键",
                    "allOf": [
                        {
                            "$ref": "#/definitions/system.LyadminDept"
                        }
                    ]
                },
                "dept_id": {
                    "description": "部门ID",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "gender": {
                    "description": "性别（男、女）",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "identity": {
                    "description": "身份（1 超级管理员,2后台用户、3前台用户）",
                    "type": "integer"
                },
                "is_active": {
                    "description": "状态(1正常、0冻结)",
                    "type": "boolean"
                },
                "is_staff": {
                    "description": "是否可登录后台",
                    "type": "boolean"
                },
                "is_superuser": {
                    "description": "是否超管",
                    "type": "boolean"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "post": {
                    "description": "岗位",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminPost"
                    }
                },
                "role": {
                    "description": "角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminRole"
                    }
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                },
                "uuid": {
                    "description": "允许读和创建",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}